Update tooling to work with 2.9 (#35)
* Update tooling to work with 2.9

**Breaking change**
With this release, the `cfg.common.allowed_ipv4_subnets` has been split into
two variables:

1. `cfg.common.allowed_ipv4_subnets.mgmt` which allows access to management
   ports tcp/22 (SSH) and tcp/9090 (Cockpit)
2. `cfg.common.allowed_ipv4_subnets.cml2` which allows access to the CML2
   application on tcp/80, tcp/443 and tcp/1122. This gets also applied to
   the PATty port range, if PATty is enabled

The default for both is `["0.0.0.0/0"]` (e.g. "any"). However, it is advisable
(and maybe even enforced by your policy) to restrict management access to a
specific subnet or a list of subnets.

- fix AWS 24.04 noble image name for the AWS mini variant (closes #32)
- fix certificate installation for 2.8 and newer in `03-letsencrypt.sh`
- mandate CML TF provider 0.8.0 or newer
- make licensing a bit more robust (`license.py`)
- additional code to install Docker related dependencies (for 2.9)
- change stabilization timer code location to match 2.8 and newer in
  `04-customize.sh`
- updated `config.yml` to include newest reference platform node definitions
  and images. Some are commented -- need to adapt before deploying to match
  what is actually needed
- updated and reformatted documentation

* making AWS work, with certs

* cleanup

* fix licensing, limit patty ports more

+++ modules/deploy/data/03-letsencrypt.sh
@@ -11,23 +11,23 @@ source /provision/vars.sh
 
 if ! is_controller; then
     echo "not a controller, exiting"
-    return
+    exit
 fi
 
-# define these in extras!
+# Define these in extras!
 # CFG_UN=""
 # CFG_PW=""
 # CFG_HN=""
 # CFG_EMAIL=""
 
-# if there's no hostname then return immediately (these scripts are sourced)
+# If there's no hostname then exit immediately
 if [ -z "${CFG_HN}" ]; then
     echo "no hostname configured, exiting"
-    return
+    exit
 fi
 
-# update our hostname on DynDNS
-IP=$(curl -s4 canhazip.com)
+# Update our hostname on DynDNS
+IP=$(curl -s4 canhazip.com) || exit
 auth=$(echo -n "$CFG_UN:$CFG_PW" | base64)
 attempts=5
 while [ $attempts -gt 0 ]; do
@@ -44,12 +44,10 @@ while [ $attempts -gt 0 ]; do
     ((attempts--))
 done
 
-echo
-
 copyfile ${CFG_HN}-fullchain.pem /tmp/fullchain.pem
 copyfile ${CFG_HN}-privkey.pem /tmp/privkey.pem
 
-if openssl x509 -text </tmp/fullchain.pem | grep ${CFG_HN}; then
+if test -f /tmp/fullchain.pem && openssl x509 -text </tmp/fullchain.pem | grep -q ${CFG_HN}; then
     mkdir -p /etc/letsencrypt/live/$CFG_HN
     mv /tmp/fullchain.pem /etc/letsencrypt/live/$CFG_HN/fullchain.pem
     mv /tmp/privkey.pem /etc/letsencrypt/live/$CFG_HN/privkey.pem
@@ -67,22 +65,21 @@ ln -s /snap/bin/certbot /usr/bin/certbot
 
 # ONLY request a cert if there's not already one present!
 if ! [ -d /etc/letsencrypt/live/$CFG_HN ]; then
+    /usr/bin/firewall-cmd --reload
     /usr/bin/certbot --domain $CFG_HN --noninteractive --nginx certonly
-    # could copy back to the cloud storage here? According to the naming
-    # scheme hostname-fullchain.pem and hostname-privkey.pem
 fi
 
-# copy the cert to the nginx configuration
-mkdir /etc/nginx/oldcerts
+# Copy the cert to the nginx configuration
+mkdir -p /etc/nginx/oldcerts
 mv /etc/nginx/*.pem /etc/nginx/oldcerts/
 cp /etc/letsencrypt/live/$CFG_HN/fullchain.pem /etc/nginx/pubkey.pem
 cp /etc/letsencrypt/live/$CFG_HN/privkey.pem /etc/nginx/privkey.pem
 
-# write the cert into the file that Cockpit uses, note that Cockpit wants
-# the cert only, not the full chain.
-cat /etc/letsencrypt/live/$CFG_HN/privkey.pem >/etc/cockpit/ws-certs.d/0-self-signed.key
-sed '/-----END CERTIFICATE-----/q' $ /etc/letsencrypt/live/$CFG_HN/fullchain.pem >/etc/cockpit/ws-certs.d/0-self-signed.cert
+# Cockpit
+rm /etc/cockpit/ws-certs.d/*
+cp /etc/letsencrypt/live/$CFG_HN/fullchain.pem /etc/cockpit/ws-certs.d/$CFG_HN.crt
+cp /etc/letsencrypt/live/$CFG_HN/privkey.pem /etc/cockpit/ws-certs.d/$CFG_HN.key
 
-# reload affected services
+# Reload affected services
 systemctl reload nginx
 systemctl restart cockpit
